Create test/support/kind_is_test.rb
